+++ includes/post-types/class-ccb-events.php
@@ -14,7 +14,7 @@
  *
  * @see   https://github.com/WebDevStudios/CPT_Core
  */
-class LO_Ccb_Events extends CPT_Core
+class LO_Ccb_Events extends LO_Ccb_Base_Post
 {
     /**
      * Bypass temp. cache
@@ -74,9 +74,6 @@ class LO_Ccb_Events extends CPT_Core
      */
     public function __construct($plugin)
     {
-        $this->plugin = $plugin;
-        $this->hooks();
-
         $page_settings = get_option('liquid_outreach_ccb_events_page_settings');
         $slug_base = !empty($page_settings['lo_events_page_permalink_base']) ? $page_settings['lo_events_page_permalink_base'] : 'outreach';
         $event_base = !empty($page_settings['lo_events_page_permalink_base_events']) ? $page_settings['lo_events_page_permalink_base_events'] : 'events';
@@ -123,97 +120,17 @@ class LO_Ccb_Events extends CPT_Core
         );
 
         $this->query_args['post_type'] = $this->post_type();
-    }
-
-    /**
-     * Initiate our hooks.
-     *
-     * @since  0.0.1
-     */
-    public function hooks()
-    {
-        add_action('cmb2_init', array($this, 'fields'));
-    }
-
-    /**
-     * Provides access to protected class properties.
-     *
-     * @since  0.2.4
-     *
-     * @param  boolean $key Specific CPT parameter to return
-     *
-     * @return mixed        Specific CPT parameter or array of singular, plural and registered name
-     */
-    public function post_type($key = 'post_type')
-    {
-        if (!$this->overrides_processed) {
-            $this->filter_values();
-        }
-
-        return parent::post_type($key);
-    }
-
-    /**
-     * Filter for overriding class properties
-     * which will be used as post arguments
-     * @since  0.2.4
-     */
-    public function filter_values()
-    {
-        if ($this->overrides_processed) {
-            return;
-        }
-
-        $args = array(
-            'singular' => $this->singular,
-            'plural' => $this->plural,
-            'post_type' => $this->post_type,
-            'arg_overrides' => $this->arg_overrides,
-        );
-
-        $filtered_args = apply_filters('lo_post_types_' . $this->id, $args, $this);
-
-        if ($filtered_args !== $args) {
-            foreach ($args as $arg => $val) {
-                if (isset($filtered_args[$arg])) {
-                    $this->{$arg} = $filtered_args[$arg];
-                }
-            }
-        }
-
-        $this->overrides_processed = true;
-    }
-
-    /**
-     * Actually registers our CPT with the merged arguments
-     * @since  0.1.0
-     */
-    public function register_post_type()
-    {
-        // Register our CPT
-        $args = register_post_type($this->post_type, $this->get_args());
-        // If error, yell about it.
-        if (is_wp_error($args)) {
-            wp_die($args->get_error_message());
-        }
-
-        // Success. Set args to what WP returns
-        $this->cpt_args = $args;
-
-        // Add this post type to our custom_post_types array
-        self::$custom_post_types[$this->post_type] = $this;
-
-        $flush_base_rewrite = get_option('lo_ccb_flush_base_rewrite');
-
-        if ($flush_base_rewrite == 'flush') {
-            flush_rewrite_rules();
-            update_option('lo_ccb_flush_base_rewrite', FALSE);
-        }
-
+    
+        $this->plugin = $plugin;
+        $this->hooks();
+    
+        add_action( 'plugins_loaded', array( $this, 'filter_values' ), 4 );
     }
 
     /**
      * Overriding get_args from parent
+     *
+     * @since 0.10.1
      * @return array
      */
     public function get_args()
@@ -231,7 +148,7 @@ class LO_Ccb_Events extends CPT_Core
             'add_new_item' => sprintf(__('Add New %s', 'cpt-core'), $this->singular),
             'edit_item' => sprintf(__('Edit %s', 'cpt-core'), $this->singular),
             'new_item' => sprintf(__('New %s', 'cpt-core'), $this->singular),
-            'all_items' => sprintf(__('All %s Events', 'cpt-core'), $this->plural),
+            'all_items' => sprintf(__('%s Events', 'cpt-core'), $this->plural),
             'view_item' => sprintf(__('View %s', 'cpt-core'), $this->singular),
             'search_items' => sprintf(__('Search %s', 'cpt-core'), $this->plural),
             'not_found' => sprintf(__('No %s', 'cpt-core'), $this->plural),
@@ -471,36 +388,6 @@ class LO_Ccb_Events extends CPT_Core
     }
 
     /**
-     * Registers admin columns to display. Hooked in via CPT_Core.
-     *
-     * @since  0.0.1
-     *
-     * @param  array $columns Array of registered column names/labels.
-     *
-     * @return array          Modified array.
-     */
-    public function columns($columns)
-    {
-        $new_column = array();
-
-        return array_merge($new_column, $columns);
-    }
-
-    /**
-     * Handles admin column display. Hooked in via CPT_Core.
-     *
-     * @since  0.0.1
-     *
-     * @param array $column Column currently being rendered.
-     * @param integer $post_id ID of post to display column for.
-     */
-    public function columns_display($column, $post_id)
-    {
-        switch ($column) {
-        }
-    }
-
-    /**
      * Retrieve lo-events.
      *
      * @since  0.2.4
@@ -664,27 +551,6 @@ class LO_Ccb_Events extends CPT_Core
         }
 
         return $events;
-    }
-
-    /**
-     * Magic getter for our object. Allows getting but not setting.
-     *
-     * @param string $field
-     *
-     * @throws Exception Throws an exception if the field is invalid.
-     * @return mixed
-     * @since  0.2.4
-     */
-    public function __get($field)
-    {
-        switch ($field) {
-            case 'id':
-            case 'arg_overrides':
-            case 'cpt_args':
-                return $this->{$field};
-            default:
-                throw new Exception('Invalid ' . __CLASS__ . ' property: ' . $field);
-        }
     }
 
 }