+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unreleased] - 0.11.4 - 13-07-2017
+## [Unreleased] - 0.11.5 - 13-07-2017
 ### Added
 - Show Images on Admin --> Outreach Categories List View
+- Add Ability to Bulk Publish Events
 
 ### Changed